Pangolin v2 with extra rewards (#77)
* Interface

* Added strategy for pangolin without extra rewarders

* add MasterChefVariableRewardsStrategy

* use local variable

* extra reward handling improvements

+++ contracts/interfaces/IMiniChefV2.sol
@@ -2,22 +2,36 @@
 pragma solidity 0.7.3;
 
 interface IMiniChefV2 {
-    function userInfo(uint pid, address user) external view returns (
-        uint256 amount,
-        uint256 rewardDebt
-    );
+    function userInfo(uint256 pid, address user) external view returns (uint256 amount, uint256 rewardDebt);
 
-    function poolInfo(uint pid) external view returns (
-        uint allocPoint,
-        uint lastRewardTime,
-        uint accRewardPerShare
-    );
+    function poolInfo(uint256 pid)
+        external
+        view
+        returns (
+            uint256 allocPoint,
+            uint256 lastRewardTime,
+            uint256 accRewardPerShare
+        );
 
-    function rewarder(uint pid) external view returns (address);
-    function lpToken(uint pid) external view returns (address);
+    function rewarder(uint256 pid) external view returns (address);
+
+    function lpToken(uint256 pid) external view returns (address);
+
     function pendingReward(uint256 _pid, address _user) external view returns (uint256);
-    function deposit(uint256 pid, uint256 amount, address to) external;
-    function withdraw(uint256 pid, uint256 amount, address to) external;
+
+    function deposit(
+        uint256 pid,
+        uint256 amount,
+        address to
+    ) external;
+
+    function withdraw(
+        uint256 pid,
+        uint256 amount,
+        address to
+    ) external;
+
     function harvest(uint256 pid, address to) external;
+
     function emergencyWithdraw(uint256 pid, address to) external;
 }
